Refactor feature to return all teas
During the refactor, I left the test mostly unchanged and decided to use a service in my controller to return the tea list. Once I did that and ran the test I got this error:
`Error creating bean with name 'teaController': Unsatisfied dependency expressed through field 'service';`

It means that its trying to inject my new service, but didn't find the bean. To fix, update the test setup by adding a `MockBean` for the new service.